requirements.txt
anyio==4.12.1
APScheduler==3.11.2
certifi==2026.1.4
charset-normalizer==3.4.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
packaging==26.0
pluggy==1.6.0
Pygments==2.19.2
pyTelegramBotAPI==4.30.0
pytest==9.0.2
python-telegram-bot[job-queue]>=21.0
Wikipedia-API>=0.6.0
pytz==2025.2
requests==2.32.5
tzlocal==5.3.1
urllib3==2.6.3
Wikipedia-API==0.9.0
